+++ packages/cli/src/zostso/issue/command/Command.handler.ts
@@ -41,7 +41,8 @@ export default class Handler extends ZosTsoBaseHandler {
         if (params.arguments.logonProcedure && params.arguments.logonProcedure !== defProc && params.arguments.suppressStartupMessages) {
             this.console.error(
                 chalk.yellow(
-                    "Warning: The logon procedure specified is not used when issuing a TSO command with the --suppress-startup-messages (--ssm) option set to true."
+                    "Warning: The logon procedure specified is not used when issuing a TSO command with the " +
+                    "--suppress-startup-messages (--ssm) option set to true."
                 )
             );
         }